requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
python-dotenv
pydantic
requests
fuzzywuzzy
python-Levenshtein
email-validator